move configuration to classes
